requirements.txt
cryptography==35.0.0
PyMySQL==1.0.2
python-dotenv==0.19.0
SQLAlchemy==1.4.25
sqlalchemy-migrate==0.13.0
SQLAlchemy-Utils==0.37.8
tornado==6.1
tornado-sqlalchemy==0.7.0